[feat/#21] feat: cities 테이블의 위도경도 컬럼을 DOUBLE형으로 변경

+++ Back/src/main/resources/schema.sql
@@ -18,8 +18,8 @@ CREATE TABLE cities (
     id INT AUTO_INCREMENT PRIMARY KEY,
     name VARCHAR(100) NOT NULL,
     image_url VARCHAR(255) NOT NULL,
-    latitude FLOAT NOT NULL,
-    longitude FLOAT NOT NULL,
+    latitude DOUBLE NOT NULL,
+    longitude DOUBLE NOT NULL,
     level INT NOT NULL DEFAULT 0
 );
